+++ src/gevent/pool.py
@@ -466,6 +466,8 @@ class Group(GroupMappingMixin):
         """
         Begin tracking the *greenlet*.
 
+        If greenlet has already completed, do nothing.
+
         If this group is :meth:`full`, then this method may block
         until it is possible to track the greenlet.
 
@@ -473,6 +475,10 @@ class Group(GroupMappingMixin):
         it is added because if this object blocks in this method,
         then the *greenlet* may run to completion before it is tracked.
         """
+        if greenlet.ready():
+            # greenlet has already finished execution
+            return
+
         try:
             rawlink = greenlet.rawlink
         except AttributeError: